requirements.txt
DistributedLock==1.2
Django==1.8.11
beautifulsoup4==4.4.1
django-annoying==0.8.7
django-bootstrap3==7.0.0
django-jsonview==0.4.3
django-m2m-history==0.3.5
django-oauth-tokens==0.6.3
django-picklefield==0.3.2
django-taggit==0.18.0
django-twitter-api==0.2.10
django-instagram==0.1.1
oauthlib==1.0.3
requests==2.9.1
requests-oauthlib==0.6.1
simplejson==3.8.2
six==1.10.0
swapper==0.1.1
tweepy==3.5.0
twitter-monitor==0.3.0

gunicorn==19.4.5
whitenoise==3.0
psycopg2==2.5.4
